Revert "Squashed commit of the following:"
This reverts commit d340a678df8a01e3fa1040250b1107ec3613bbab.

+++ Client/mods/deathmatch/logic/CResourceFileDownloadManager.cpp
@@ -154,7 +154,7 @@ void CResourceFileDownloadManager::DoPulse()
     // Call Lua event 'onClientTransferBoxProgressChange'
     CLuaArguments arguments;
     arguments.PushNumber(uiDownloadedSizeTotal);
-    arguments.PushNumber(static_cast<double>(GetTransferBox()->GetDownloadTotalSize()));
+    arguments.PushNumber(GetTransferBox()->GetDownloadTotalSize());
 
     g_pClientGame->GetRootEntity()->CallEvent("onClientTransferBoxProgressChange", arguments, false);
 
@@ -290,7 +290,8 @@ void CResourceFileDownloadManager::StaticDownloadFinished(const SHttpDownloadRes
 ///////////////////////////////////////////////////////////////
 void CResourceFileDownloadManager::DownloadFinished(const SHttpDownloadResult& result)
 {
-    auto iter = std::find(std::begin(m_ActiveFileDownloadList), std::end(m_ActiveFileDownloadList), reinterpret_cast<CDownloadableResource*>(result.pObj));
+    auto iter = std::find(std::begin(m_ActiveFileDownloadList), std::end(m_ActiveFileDownloadList),
+                          reinterpret_cast<CDownloadableResource*>(result.pObj));
 
     if (iter == std::end(m_ActiveFileDownloadList))
         return;
