+++ Server/mods/deathmatch/logic/luadefs/CLuaPedDefs.cpp
@@ -10,9 +10,6 @@
  *****************************************************************************/
 
 #include "StdInc.h"
-#include "CLuaPedDefs.h"
-#include "CStaticFunctionDefinitions.h"
-#include "CScriptArgReader.h"
 
 void CLuaPedDefs::LoadFunctions()
 {
@@ -34,7 +31,7 @@ void CLuaPedDefs::LoadFunctions()
         {"getPedTotalAmmo", GetPedTotalAmmo},
         {"getPedWeapon", GetPedWeapon},
         {"getPedClothes", GetPedClothes},
-        {"isPedWearingJetpack", DoesPedHaveJetPack},            // introduced in 1.5.5-9.13846
+        {"isPedWearingJetpack", DoesPedHaveJetPack}, // introduced in 1.5.5-9.13846
         {"isPedOnGround", IsPedOnGround},
         {"getPedFightingStyle", GetPedFightingStyle},
         {"getPedWalkingStyle", GetPedMoveAnim},
@@ -56,7 +53,7 @@ void CLuaPedDefs::LoadFunctions()
         {"setPedStat", SetPedStat},
         {"addPedClothes", AddPedClothes},
         {"removePedClothes", RemovePedClothes},
-        {"setPedWearingJetpack", SetPedWearingJetpack},            // introduced in 1.5.5-9.13846
+        {"setPedWearingJetpack", SetPedWearingJetpack}, // introduced in 1.5.5-9.13846
         {"setPedFightingStyle", SetPedFightingStyle},
         {"setPedWalkingStyle", SetPedMoveAnim},
         {"setPedGravity", SetPedGravity},
@@ -179,7 +176,7 @@ int CLuaPedDefs::GetValidPedModels(lua_State* luaVM)
 {
     int iIndex = 0;
     lua_newtable(luaVM);
-    for (unsigned short i = 0; i <= 312; i++)
+    for (int i = 0; i <= 312; i++)
     {
         if (CPlayerManager::IsValidPlayerModel(i))
         {
@@ -442,8 +439,7 @@ int CLuaPedDefs::SetPedAnimation(lua_State* luaVM)
         szBlock = strBlockName.empty() ? NULL : strBlockName.c_str();
         szAnim = strAnimName.empty() ? NULL : strAnimName.c_str();
 
-        if (CStaticFunctionDefinitions::SetPedAnimation(pPed, szBlock, szAnim, iTime, iBlend, bLoop, bUpdatePosition, bInterruptable, bFreezeLastFrame,
-                                                        bTaskToBeRestoredOnAnimEnd))
+        if (CStaticFunctionDefinitions::SetPedAnimation(pPed, szBlock, szAnim, iTime, iBlend, bLoop, bUpdatePosition, bInterruptable, bFreezeLastFrame, bTaskToBeRestoredOnAnimEnd))
         {
             lua_pushboolean(luaVM, true);
             return 1;
